Add SearchEngineResultToRecordsConverter
